+++ apps/shelter/src/apis/shelter.ts
@@ -1,18 +1,4 @@
-import axiosInstance from '../axiosInstance';
-
-type SignUpParams = {
-  email: string;
-  password: string;
-  name: string;
-  address: string;
-  addressDetail: string;
-  phoneNumber: string;
-  sparePhoneNumber: string;
-  isOpenedAddress: boolean;
-};
-
-export const signUpShelter = (signUpParams: SignUpParams) =>
-  axiosInstance.post<unknown, SignUpParams>('/shelters', signUpParams);
+import axiosInstance from 'shared/apis/axiosInstance';
 
 type ShelterInfo = {
   name: string;
@@ -24,8 +10,10 @@ type ShelterInfo = {
   isOpenedAddress: boolean;
 };
 
-export const updateShelterInfo = (shelterInfo: ShelterInfo) =>
-  axiosInstance.patch<unknown, ShelterInfo>('/shelters/me', shelterInfo);
+type PasswordUpdateParams = {
+  newPassword: string;
+  oldPassword: string;
+};
 
 export const getShelterInfo = () =>
   axiosInstance.get<{
@@ -39,10 +27,8 @@ export const getShelterInfo = () =>
     shelterIsOpenedAddress: boolean;
   }>('/shelters/me');
 
-type PasswordUpdateParams = {
-  newPassword: string;
-  oldPassword: string;
-};
+export const updateShelterInfo = (shelterInfo: ShelterInfo) =>
+  axiosInstance.patch<unknown, ShelterInfo>('/shelters/me', shelterInfo);
 
 export const updatePassword = (passwordUpdateParams: PasswordUpdateParams) =>
   axiosInstance.patch<unknown, PasswordUpdateParams>(
@@ -58,11 +44,29 @@ export const updateAddressStatus = (isOpenedAddress: boolean) =>
     }
   >('/shelters/me/address/status', { isOpenedAddress });
 
-export const checkDuplicatedShelterEmail = (email: string) => {
-  return axiosInstance.post<{ isDuplicated: boolean }, { email: string }>(
-    '/shelters/email',
-    {
-      email,
+export const getShelterReviewList = (
+  shelterId: number,
+  pageNumber: number,
+  pageSize: number,
+) =>
+  axiosInstance.get<{
+    pageInfo: {
+      totalElements: number;
+      hasNext: boolean;
+    };
+    reviews: {
+      reviewId: number;
+      reviewCreatedAt: string;
+      reviewContent: string;
+      reviewImageUrls: string[];
+      volunteerName: string;
+      volunteerTemperature: number;
+      volunteerReviewCount: number;
+      volunteerImageUrl: number;
+    }[];
+  }>(`/shelters/${shelterId}/reviews`, {
+    params: {
+      pageSize,
+      pageNumber,
     },
-  );
-};
+  });
